+++ CHANGELOG.md
@@ -6,7 +6,8 @@ PDF417 Changelog
 
 * Added validation of options in renderers.
 * Fixed a bug in ImageRenderer where padding was not in bgColor, but white.
-* Upgraded Intrevention/Image to v2
+* Upgraded Intrevention/Image to v2.
+* Added 'quality' option to ImageRenderer (used for JPG only).
 
 0.1.0 (2014-12-24)
 ------------------